feat: enhance caching system and update fetch options
- Introduced new cache options in `FetchOptions`, including `cache`, `cacheKey`, `cacheTTL`, `cacheNamespace`, `cacheBaseUrl`, and `noCache`.
- Updated the `createCache` function to support dynamic imports for the FileSystemCache, optimizing for different environments.
- Improved cache key generation and validation methods for better performance and reliability.
- Refactored `fetchFromWebWorker` and `fetchFromWeb` to utilize the enhanced caching capabilities.

These changes improve the flexibility and efficiency of the caching system, allowing for better management of cached content across different environments.
